No setPrototypeOf due to poor performance in Node 12.16.1+, req properties are all methods now
