+++ lib/middleware/init.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const { route } = require('../application')
-var setPrototypeOf = require('setprototypeof')
 
 function defineGetter(obj, name, getter, setter) {
   Object.defineProperty(obj, name, {
@@ -52,9 +51,35 @@ var init = function (options, router, slave) {
 
     req.rData_internal.initDone = true
 
+    Object.assign(req, req.app.request.__proto__)
+    Object.assign(req, req.app.request)
+
+    Object.assign(res, req.app.response.__proto__)
+    Object.assign(res, req.app.response)
+
+    function currentUrl () {
+      const route = this.route
+      if (route.middleware) {
+        let url = this.url.replace(route.pattern, '')
+        url = url || '/'
+        return url
+      }
+      return this.url
+    }
     
-    setPrototypeOf(req, req.app.request)
-    setPrototypeOf(res, req.app.response)
+    defineGetter(req, 'currentUrl', currentUrl)
+    
+    function baseUrl () {
+      const curRouterMountpath = this.rData_internal.tree[this.rData_internal.tree.length - 1]
+      const baseUrl = curRouterMountpath && curRouterMountpath !== '/' ? this.originalUrl.replace(this.currentUrl, '') : ''
+      return baseUrl
+    }
+    
+    function baseUrlSetter () {
+      return this
+    }
+    
+    defineGetter(req, 'baseUrl', baseUrl, baseUrlSetter)
     
 
     function method () {
@@ -84,8 +109,6 @@ var init = function (options, router, slave) {
     }
 
     defineGetter(req, 'url', urlGetter, urlSetter)
-
-    // these properties need to be redefined from here
 
     return next()
   }
