+++ lib/request.js
@@ -48,7 +48,7 @@ req.header = function header (name) {
   }
 }
 
-req.accepts = function () {
+req.accepts = function reqAccepts () {
   var accept = accepts(this)
   return accept.types.apply(accept, arguments)
 }
@@ -114,7 +114,7 @@ req.is = function is (types) {
   return typeis(this, arr)
 }
 
-defineGetter(req, 'protocol', function protocol () {
+req.protocol = function protocol () {
   var proto = this.connection.encrypted
     ? 'https'
     : 'http'
@@ -132,22 +132,18 @@ defineGetter(req, 'protocol', function protocol () {
   return index !== -1
     ? header.substring(0, index).trim()
     : header.trim()
-})
-
-function secure () {
-  return this.protocol === 'https'
 }
 
-defineGetter(req, 'secure', secure)
+req.secure = function secure () {
+  return this.protocol() === 'https'
+}
 
-function ip () {
+req.ip = function ip () {
   var trust = this.app.get('trust proxy fn')
   return proxyaddr(this, trust)
 }
 
-defineGetter(req, 'ip', ip)
-
-function ips () {
+req.ips = function ips () {
   var trust = this.app.get('trust proxy fn')
   var addrs = proxyaddr.all(this, trust)
 
@@ -158,10 +154,8 @@ function ips () {
   return addrs
 }
 
-defineGetter(req, 'ips', ips)
-
-function subdomains () {
-  var hostname = this.hostname
+req.subdomains = function subdomains () {
+  var hostname = this.hostname()
   
   if (!hostname) return []
 
@@ -172,32 +166,6 @@ function subdomains () {
 
   return subdomains.slice(offset)
 }
-
-defineGetter(req, 'subdomains', subdomains)
-
-function currentUrl () {
-  const route = this.route
-  if (route.middleware) {
-    let url = this.url.replace(route.pattern, '')
-    url = url || '/'
-    return url
-  }
-  return this.url
-}
-
-defineGetter(req, 'currentUrl', currentUrl)
-
-function baseUrl () {
-  const curRouterMountpath = this.rData_internal.tree[this.rData_internal.tree.length - 1]
-  const baseUrl = curRouterMountpath && curRouterMountpath !== '/' ? this.originalUrl.replace(this.currentUrl, '') : ''
-  return baseUrl
-}
-
-function baseUrlSetter () {
-  return this
-}
-
-defineGetter(req, 'baseUrl', baseUrl, baseUrlSetter)
 
 
 req.setUrl = function setUrl (url) {
@@ -210,7 +178,7 @@ req.setMethod = function setMethod (method) {
   return this
 }
 
-function hostname () {
+req.hostname = function hostname () {
   var trust = this.app.get('trust proxy fn')
   var host = this.headers['x-forwarded-host']
 
@@ -235,17 +203,13 @@ function hostname () {
     : host
 }
 
-defineGetter(req, 'hostname', hostname)
-
 // TODO: change req.host to return host in next major
 
-var host = deprecate.function(function host () {
-  return this.hostname
+req.host = deprecate.function(function host () {
+  return this.hostname()
 }, 'req.host: Use req.hostname instead')
 
-defineGetter(req, 'host', host)
-
-function reqFresh () {
+req.fresh = function reqFresh () {
   var method = this.rData_internal.method
   var res = this.res
   var status = res.statusCode
@@ -264,26 +228,11 @@ function reqFresh () {
   return false
 }
 
-defineGetter(req, 'fresh', reqFresh)
-
-function stale () {
-  return !this.fresh
+req.stale = function stale () {
+  return !this.fresh()
 }
 
-defineGetter(req, 'stale', stale)
-
-function xhr () {
+req.xhr = function xhr () {
   var val = this.headers['x-requested-with'] || ''
   return val.toLowerCase() === 'xmlhttprequest'
-}
-
-defineGetter(req, 'xhr', xhr)
-
-function defineGetter(obj, name, getter, setter) {
-  Object.defineProperty(obj, name, {
-    configurable: true,
-    enumerable: true,
-    get: getter,
-    set: setter
-  });
 }
