+++ lib/response.js
@@ -169,7 +169,7 @@ function resSend (res, chunk) { // split to avoid deprecation checks when return
   }
 
   // freshness
-  const fresh = req.fresh
+  const fresh = req.fresh()
 
   // strip irrelevant headers
   if ((fresh && (res.statusCode = 304)) || res.statusCode === 204 || res.statusCode === 304) {
